requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
web3==7.6.0
python-bitcoinlib==0.12.2
requests==2.32.3
gunicorn==23.0.0
python-decouple==3.8
Pillow==11.1.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.52
drf-spectacular==0.28.0
django-extensions==3.2.3
whitenoise==6.8.2
